Absorb ssplit-cpp and clean headers towards packaging
Change simplifies submodule dependencies towards packaging for
distribution. Path being taken is:

1. `google/sentencepiece` instead of `browsermt/sentencepiece`, which
should correspond to `libsentencepiece-dev`. There are some disparities
in training code, but should be okay for inference. It appears it is not
possible to rely on `libsentencepiece-dev` or equivalents because these
do no export `SentencePieceText` (see relevant SentencePieceAPI docs).
2. `browsermt/ssplit-cpp` is absorbed here into
`{Splitter,Regex}.{hh,cc}`.
3. 8-bit GEMM provider is enabled using `mozilla/gemmology`, which using
CMake is pointed to for include-directory. Since this is header-only and
used only in `QMM.cc`, gemmology headers end up private and is not
needed to distribute. This adds an `xsimd` dependency, which appears to
be available via repositories.

Links

- SentencePiece API
https://github.com/google/sentencepiece/blob/8cbdf13794284c30877936f91c6f31e2c1d5aef7/src/sentencepiece_processor.h#L387-L401
- google/sentencepiece https://github.com/google/sentencepiece
- browsermt/sentencepiece https://github.com/browsermt/sentencepiece
- browsermt/ssplit-cpp https://github.com/browsermt/ssplit-cpp
- mozilla/gemmology https://github.com/mozilla/gemmology
- xsimd https://github.com/xtensor-stack/xsimd
- libsentencepiece-dev
https://packages.debian.org/stable/libdevel/libsentencepiece-dev

Pull-Request: https://github.com/jerinphilip/slimt/pull/13