Miscellaneous improvements
The pybind11 linking to shared executable caused me conflicts (virtual
environment versus system). So as to not trip over this again, and to
also allow seamless PyPI packaging, considering linking from pybind11
statically to `libslimt.a` instead of shared to `libslimt.so`).

The config preset methods are now prefixed with `SLIMT_EXPORT`. 

Moved `Encoding` to `slimt/Types.hh` which is not necessary, but not
harmful either and possibly okay to keep.

Broken package-management which used the extension of the tar 
archive is now fixed, and there's possible room for refinement.

Pull Request: https://github.com/jerinphilip/slimt/pull/41